requirements.txt
Django==1.6.5
django-toolbelt==0.0.1
python-social-auth==0.1.26
